+++ web/src/Components/ScrobbleTable.js
@@ -3,37 +3,32 @@ import { Link } from 'react-router-dom';
 
 const ScrobbleTable = (props) => {
     return (
-        <div style={{width: `100%`, maxWidth: `900px`}}>
-        <table style={{width: `100%`}} border={1} cellPadding={5}>
-           <thead>
-              <tr>
-                <td>Timestamp</td>
-                <td>Track</td>
-                <td>Artist</td>
-                <td>Album</td>
-              </tr>
-           </thead>
-           <tbody>
-              {
-                  props.data &&
-                  props.data.map(function (element) {
-                     let localTime = new Date(element.time);
-                     return <tr key={element.uuid}>
-                        <td>{localTime.toLocaleString()}</td>
-                        <td>
-                           <Link
-                               key={element.track.uuid}
-                              to={"/track/"+element.track.uuid}
-                               >{element.track.name}
-                           </Link>
-                        </td>
-                        <td>{element.artist}</td>
-                        <td>{element.album}</td>
-                     </tr>;
-                  })
-              }
-           </tbody>
-        </table>
+        <div style={{
+         border: `1px solid #FFFFFF`,
+         width: `100%`,
+         display: `flex`,
+         flexWrap: `wrap`,
+         minWidth: `300px`,
+         maxWidth: `900px`,
+        }}>
+         {
+            props.data &&
+            props.data.map(function (element) {
+               let localTime = new Date(element.time);
+               return <div style={{borderBottom: `1px solid #CCC`, width: `100%`, padding: `2px`}} key={"box" + element.time}>
+                    {localTime.toLocaleString()}<br/>
+                    <Link
+                        key={"artist" + element.time}
+                        to={"/artist/"+element.artist.uuid}
+                     >{element.artist.name}</Link> -
+                     <Link
+                        key={"track" + element.time}
+                        to={"/track/"+element.track.uuid}
+                     > {element.track.name}</Link>
+               </div>;
+
+            })
+         }
       </div>
     );
 }