+++ web/src/Pages/Album.js
@@ -49,8 +49,8 @@ const Album = (route) => {
         {album.name}
       </h1>
       <div className="pageBody">
-        MusicBrainzId: {album.mbid}<br/>
-        SpotifyID: {album.spotify_id}
+        {album.mbid && <a rel="noreferrer" target="_blank" href={"https://musicbrainz.org/album/" + album.mbid}>Open on MusicBrainz<br/></a>}
+        {album.spotify_id && <a rel="noreferrer" target="_blank" href={"https://open.spotify.com/album/" + album.spotify_id}>Open on Spotify<br/></a>}
       </div>
     </div>
   );