+++ web/src/Pages/Artist.js
@@ -48,9 +48,10 @@ const Artist = (route) => {
       <h1>
         {artist.name}
       </h1>
-      <div className="pageBody">
-        MusicBrainzId: {artist.mbid}<br/>
-        SpotifyID: {artist.spotify_id}
+      <div className="pageBody" style={{textAlign: `center`}}>
+        <img src={artist.img} alt={artist.name} style={{maxWidth: `300px`, maxHeight: `300px`}}/><br/><br/>
+        {artist.mbid && <a rel="noreferrer" target="_blank" href={"https://musicbrainz.org/artist/" + artist.mbid}>Open on MusicBrainz<br/></a>}
+        {artist.spotify_id && <a rel="noreferrer" target="_blank" href={"https://open.spotify.com/artist/" + artist.spotify_id}>Open on Spotify<br/></a>}
       </div>
     </div>
   );